Big fixes for chronoshift exploits.
The chronoshift previously would give its user a lot of details about the units underneath its cursor even if the user did not have direct visibility of the area. The functionality is now as follows:

- If a user has no exploration of an area, chronoshift is not allowed.
- If a user has explored an area but has no visibility, they can chronoshift as much as they please from the area, but will not receive any UI feedback if there is a unit present.
- If a user has GPS, the cursor will change color appropriately both during targeting and activation, so the user can use the cursor to see when units move in and out of the targeted space.
- If the user has full visibility of the area, the cursor will also include unit previews when placing the units.

This commit also fixes a bug that checked the target unit's exploration for determining valid placement instead of the chronosphere owner's.

Fixes #7 and fixes #10.